+++ src/Helpers/Commands.ts
@@ -1,22 +1,25 @@
 export class Commands {
-    public static PortalList(){
+    public static PortalList() {
         return 'pac paportal list';
     }
 
-    public static DownloadPortal(localPath:string, websiteId: string, overwritePortal: string | undefined){
-        let overwriteText = overwritePortal === 'Yes' ? `-o true`: '';
+    public static DownloadPortal(localPath: string, websiteId: string, overwritePortal: string | undefined) {
+        let overwriteText = overwritePortal === 'Yes' ? `-o true` : '';
         return `pac paportal download -p "${localPath}" -id ${websiteId} ${overwriteText}`;
     }
 
-    public static UploadPortal(localPath:string){
-        return `pac paportal upload -p "${localPath}"`;
+    public static UploadPortal(localPath: string, deploymentProfile: string | undefined) {
+        const command = deploymentProfile
+            ? `pac paportal upload -p "${localPath}" --deploymentProfile "${deploymentProfile}"`
+            : `pac paportal upload -p "${localPath}"`;
+        return command;
     }
 
-    public static AuthList(){
+    public static AuthList() {
         return 'pac auth list';
     }
 
-    public static AuthCreateProfile(dataverseUrl:string){
+    public static AuthCreateProfile(dataverseUrl: string) {
         return `pac auth create -u ${dataverseUrl}`;
     }